refactor-#14: made all the suggested changes
