+++ frontend/src/components/BrowseByCategory.tsx
@@ -2,263 +2,19 @@
 import React, { useState } from "react";
 import { Card, CardContent } from "./ui/card";
 import { motion, AnimatePresence } from "framer-motion";
-
-// Star Rating Component
-const StarRating = ({ rating }: { rating: number }) => {
-  const fullStars = Math.floor(rating);
-  const hasHalfStar = rating % 1 !== 0;
-  const emptyStars = 5 - fullStars - (hasHalfStar ? 1 : 0);
-
-  return (
-    <div className="flex items-center gap-1">
-      {/* Full stars */}
-      {Array.from({ length: fullStars }, (_, i) => (
-        <motion.svg
-          key={`full-${i}`}
-          className="h-4 w-4 text-yellow-400"
-          fill="currentColor"
-          viewBox="0 0 20 20"
-          initial={{ opacity: 0, scale: 0 }}
-          animate={{ opacity: 1, scale: 1 }}
-          transition={{ delay: i * 0.1, type: "spring", stiffness: 200 }}
-        >
-          <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
-        </motion.svg>
-      ))}
-
-      {/* Half star */}
-      {hasHalfStar && (
-        <motion.div
-          className="relative"
-          initial={{ opacity: 0, scale: 0 }}
-          animate={{ opacity: 1, scale: 1 }}
-          transition={{
-            delay: fullStars * 0.1,
-            type: "spring",
-            stiffness: 200,
-          }}
-        >
-          <svg
-            className="h-4 w-4 text-gray-300"
-            fill="currentColor"
-            viewBox="0 0 20 20"
-          >
-            <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
-          </svg>
-          <svg
-            className="absolute top-0 left-0 h-4 w-4 text-yellow-400"
-            fill="currentColor"
-            viewBox="0 0 20 20"
-            style={{ clipPath: "inset(0 50% 0 0)" }}
-          >
-            <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
-          </svg>
-        </motion.div>
-      )}
-
-      {/* Empty stars */}
-      {Array.from({ length: emptyStars }, (_, i) => (
-        <motion.svg
-          key={`empty-${i}`}
-          className="h-4 w-4 text-gray-300"
-          fill="currentColor"
-          viewBox="0 0 20 20"
-          initial={{ opacity: 0, scale: 0 }}
-          animate={{ opacity: 1, scale: 1 }}
-          transition={{
-            delay: (fullStars + (hasHalfStar ? 1 : 0) + i) * 0.1,
-            type: "spring",
-            stiffness: 200,
-          }}
-        >
-          <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
-        </motion.svg>
-      ))}
-
-      <span className="text-muted-foreground ml-1 text-sm">
-        {rating.toFixed(1)}
-      </span>
-    </div>
-  );
-};
-const categories = {
-  "Graphics & Design": {
-    subcategories: [
-      "Logo Design",
-      "Brand Style Guides",
-      "Game Art",
-      "Illustration",
-      "Web & Mobile Design",
-      "Social Media Design",
-    ],
-    icon: "🎨",
-    color: "from-primary to-primary/70",
-    rating: 4.8,
-    skillsCount: 1247,
-  },
-  "Digital Marketing": {
-    subcategories: [
-      "Social Media Marketing",
-      "SEO",
-      "Content Marketing",
-      "Email Marketing",
-      "Video Marketing",
-      "Influencer Marketing",
-    ],
-    icon: "📱",
-    color: "from-secondary to-secondary/70",
-    rating: 4.7,
-    skillsCount: 892,
-  },
-  "Writing & Translation": {
-    subcategories: [
-      "Articles & Blog Posts",
-      "Translation",
-      "Proofreading & Editing",
-      "Resume Writing",
-      "Technical Writing",
-      "Creative Writing",
-    ],
-    icon: "✍️",
-    color: "from-accent to-accent/70",
-    rating: 4.6,
-    skillsCount: 634,
-  },
-  "Video & Animation": {
-    subcategories: [
-      "Video Editing",
-      "Short Video Ads",
-      "Animated Explainers",
-      "Logo Animation",
-      "Intros & Outros",
-      "Character Animation",
-    ],
-    icon: "🎬",
-    color: "from-destructive to-destructive/70",
-    rating: 4.9,
-    skillsCount: 759,
-  },
-  "Music & Audio": {
-    subcategories: [
-      "Voice Over",
-      "Mixing & Mastering",
-      "Podcast Editing",
-      "Music Production",
-      "Audiobook Production",
-    ],
-    icon: "🎵",
-    color: "from-primary to-secondary",
-    rating: 4.5,
-    skillsCount: 423,
-  },
-  "Programming & Tech": {
-    subcategories: [
-      "Website Development",
-      "Mobile App Development",
-      "E-Commerce Development",
-      "WordPress",
-      "Game Development",
-      "Cybersecurity",
-    ],
-    icon: "💻",
-    color: "from-foreground to-foreground/70",
-    rating: 4.8,
-    skillsCount: 1563,
-  },
-  Business: {
-    subcategories: [
-      "Virtual Assistant",
-      "Data Entry",
-      "Market Research",
-      "Business Plans",
-      "Project Management",
-    ],
-    icon: "💼",
-    color: "from-muted-foreground to-muted-foreground/70",
-    rating: 4.4,
-    skillsCount: 567,
-  },
-  Lifestyle: {
-    subcategories: [
-      "Online Tutoring",
-      "Fitness Lessons",
-      "Cooking Lessons",
-      "Life Coaching",
-      "Travel Planning",
-    ],
-    icon: "🌟",
-    color: "from-accent to-primary",
-    rating: 4.3,
-    skillsCount: 389,
-  },
-  "Data Science & Analytics": {
-    subcategories: [
-      "Data Analysis",
-      "Data Visualization",
-      "Machine Learning",
-      "Data Engineering",
-      "Data Mining",
-      "Statistical Analysis",
-    ],
-    icon: "📊",
-    color: "from-secondary to-accent",
-    rating: 4.7,
-    skillsCount: 892,
-  },
-};
-
-// Animation variants for container
-const containerVariants = {
-  hidden: { opacity: 0 },
-  visible: {
-    opacity: 1,
-    transition: {
-      staggerChildren: 0.1,
-      delayChildren: 0.2,
-    },
-  },
-};
-
-// Animation variants for individual cards
-const cardVariants = {
-  hidden: {
-    opacity: 0,
-    y: 50,
-    scale: 0.9,
-  },
-  visible: {
-    opacity: 1,
-    y: 0,
-    scale: 1,
-    transition: {
-      type: "spring" as const,
-      damping: 25,
-      stiffness: 120,
-      duration: 0.8,
-    },
-  },
-};
-
-// Hover animation variants
-const hoverVariants = {
-  hover: {
-    scale: 1.05,
-    y: -10,
-    transition: {
-      type: "spring" as const,
-      damping: 20,
-      stiffness: 300,
-    },
-  },
-  tap: {
-    scale: 0.98,
-    transition: {
-      type: "spring" as const,
-      damping: 20,
-      stiffness: 400,
-    },
-  },
-};
+import { X, ChevronRight } from "lucide-react";
+import { categories } from "@/data/constants";
+import { StarRating } from "./ui/star-rating";
+import {
+  containerVariants,
+  cardVariants,
+  hoverVariants,
+  wordAnimationVariants,
+  iconAnimationVariants,
+  modalVariants,
+  buttonAnimationVariants,
+  effectVariants,
+} from "@/lib/animations";
 
 const BrowseByCategory = () => {
   const [selectedCategory, setSelectedCategory] = useState<string | null>(null);
@@ -303,17 +59,9 @@ const BrowseByCategory = () => {
             <motion.span
               key={index}
               className="from-foreground via-primary to-foreground mr-3 inline-block bg-gradient-to-r bg-clip-text text-transparent"
-              whileHover={{
-                scale: 1.1,
-                y: -5,
-                transition: {
-                  type: "spring",
-                  stiffness: 300,
-                  damping: 15,
-                },
-              }}
-              initial={{ opacity: 0, y: 20 }}
-              animate={{ opacity: 1, y: 0 }}
+              whileHover={wordAnimationVariants.title.whileHover}
+              initial={wordAnimationVariants.title.initial}
+              animate={wordAnimationVariants.title.animate}
               transition={{
                 delay: 0.3 + index * 0.1,
                 duration: 0.6,
@@ -338,16 +86,9 @@ const BrowseByCategory = () => {
               <motion.span
                 key={index}
                 className="mr-2 inline-block"
-                whileHover={{
-                  scale: 1.05,
-                  transition: {
-                    type: "spring",
-                    stiffness: 400,
-                    damping: 10,
-                  },
-                }}
-                initial={{ opacity: 0, y: 15 }}
-                animate={{ opacity: 1, y: 0 }}
+                whileHover={wordAnimationVariants.subtitle.whileHover}
+                initial={wordAnimationVariants.subtitle.initial}
+                animate={wordAnimationVariants.subtitle.animate}
                 transition={{
                   delay: 0.6 + index * 0.05,
                   duration: 0.5,
@@ -443,13 +184,9 @@ const BrowseByCategory = () => {
                           left: `${20 + i * 30}%`,
                           top: `${10 + i * 20}%`,
                         }}
-                        animate={{
-                          y: [0, -10, 0],
-                          opacity: [0.3, 1, 0.3],
-                        }}
+                        animate={effectVariants.particles.animate}
                         transition={{
-                          duration: 2,
-                          repeat: Infinity,
+                          ...effectVariants.particles.transition,
                           delay: i * 0.2,
                         }}
                       />
@@ -459,10 +196,10 @@ const BrowseByCategory = () => {
                   {/* Shimmer effect - Now covers the entire card */}
                   <motion.div
                     className="via-primary/10 pointer-events-none absolute inset-0 -translate-x-full bg-gradient-to-r from-transparent to-transparent transition-transform duration-1000 group-hover:translate-x-full group-active:translate-x-full"
-                    initial={{ x: "-100%" }}
-                    whileHover={{ x: "100%" }}
-                    whileTap={{ x: "100%" }}
-                    transition={{ duration: 0.8 }}
+                    initial={effectVariants.shimmer.initial}
+                    whileHover={effectVariants.shimmer.whileHover}
+                    whileTap={effectVariants.shimmer.whileTap}
+                    transition={effectVariants.shimmer.transition}
                   />
 
                   <CardContent className="relative z-10 flex h-full flex-col p-6">
@@ -479,24 +216,8 @@ const BrowseByCategory = () => {
                           damping: 10,
                           stiffness: 400,
                         }}
-                        whileHover={{
-                          scale: 1.1,
-                          rotate: 12,
-                          transition: {
-                            type: "spring",
-                            damping: 10,
-                            stiffness: 400,
-                          },
-                        }}
-                        whileTap={{
-                          scale: 1.05,
-                          rotate: 8,
-                          transition: {
-                            type: "spring",
-                            damping: 10,
-                            stiffness: 400,
-                          },
-                        }}
+                        whileHover={iconAnimationVariants.whileHover}
+                        whileTap={iconAnimationVariants.whileTap}
                       >
                         {icon}
                       </motion.span>
@@ -539,8 +260,8 @@ const BrowseByCategory = () => {
                         transition={{ delay: 0.5 + index * 0.1 }}
                       >
                         <motion.div
-                          whileHover={{ scale: 1.05 }}
-                          whileTap={{ scale: 0.95 }}
+                          whileHover={buttonAnimationVariants.whileHover}
+                          whileTap={buttonAnimationVariants.whileTap}
                         >
                           <StarRating rating={rating} />
                         </motion.div>
@@ -564,10 +285,10 @@ const BrowseByCategory = () => {
                     {/* Hover indicator */}
                     <motion.div
                       className="from-primary to-primary/50 absolute right-0 bottom-0 left-0 h-1 bg-gradient-to-r opacity-0 transition-opacity duration-300 group-hover:opacity-100 group-active:opacity-100"
-                      initial={{ scaleX: 0 }}
-                      whileHover={{ scaleX: 1 }}
-                      whileTap={{ scaleX: 1 }}
-                      transition={{ type: "spring", stiffness: 400 }}
+                      initial={effectVariants.hoverIndicator.initial}
+                      whileHover={effectVariants.hoverIndicator.whileHover}
+                      whileTap={effectVariants.hoverIndicator.whileTap}
+                      transition={effectVariants.hoverIndicator.transition}
                     />
                   </CardContent>
                 </Card>
@@ -582,17 +303,17 @@ const BrowseByCategory = () => {
         {isModalOpen && selectedCategory && (
           <motion.div
             className="fixed inset-0 z-50 flex items-center justify-center bg-black/50 p-4"
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
+            initial={modalVariants.overlay.initial}
+            animate={modalVariants.overlay.animate}
+            exit={modalVariants.overlay.exit}
             onClick={handleCloseModal}
           >
             <motion.div
               className="bg-background max-h-[80vh] w-full max-w-2xl overflow-hidden rounded-xl border shadow-2xl"
-              initial={{ scale: 0.8, opacity: 0, y: 20 }}
-              animate={{ scale: 1, opacity: 1, y: 0 }}
-              exit={{ scale: 0.8, opacity: 0, y: 20 }}
-              transition={{ type: "spring", damping: 25, stiffness: 300 }}
+              initial={modalVariants.content.initial}
+              animate={modalVariants.content.animate}
+              exit={modalVariants.content.exit}
+              transition={modalVariants.content.transition}
               onClick={(e) => e.stopPropagation()}
             >
               {/* Modal Header */}
@@ -615,22 +336,10 @@ const BrowseByCategory = () => {
                   <motion.button
                     className="hover:bg-secondary rounded-lg p-2 transition-colors"
                     onClick={handleCloseModal}
-                    whileHover={{ scale: 1.1 }}
-                    whileTap={{ scale: 0.95 }}
+                    whileHover={buttonAnimationVariants.whileHover}
+                    whileTap={buttonAnimationVariants.whileTap}
                   >
-                    <svg
-                      className="h-6 w-6"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M6 18L18 6M6 6l12 12"
-                      />
-                    </svg>
+                    <X className="h-6 w-6" />
                   </motion.button>
                 </div>
               </div>
@@ -660,19 +369,7 @@ const BrowseByCategory = () => {
                           initial={{ x: -10 }}
                           whileHover={{ x: 0 }}
                         >
-                          <svg
-                            className="text-primary h-5 w-5"
-                            fill="none"
-                            stroke="currentColor"
-                            viewBox="0 0 24 24"
-                          >
-                            <path
-                              strokeLinecap="round"
-                              strokeLinejoin="round"
-                              strokeWidth={2}
-                              d="M9 5l7 7-7 7"
-                            />
-                          </svg>
+                          <ChevronRight className="text-primary h-5 w-5" />
                         </motion.div>
                       </div>
                     </motion.button>
@@ -693,8 +390,8 @@ const BrowseByCategory = () => {
                   <motion.button
                     className="bg-primary text-primary-foreground hover:bg-primary/90 rounded-lg px-4 py-2 transition-colors"
                     onClick={handleCloseModal}
-                    whileHover={{ scale: 1.05 }}
-                    whileTap={{ scale: 0.95 }}
+                    whileHover={buttonAnimationVariants.whileHover}
+                    whileTap={buttonAnimationVariants.whileTap}
                   >
                     Close
                   </motion.button>
