+++ frontend/src/components/ui/star-rating.tsx
@@ -1,0 +1,84 @@
+"use client";
+import { motion } from "framer-motion";
+
+interface StarRatingProps {
+  rating: number;
+}
+
+export const StarRating = ({ rating }: StarRatingProps) => {
+  const fullStars = Math.floor(rating);
+  const hasHalfStar = rating % 1 !== 0;
+  const emptyStars = 5 - fullStars - (hasHalfStar ? 1 : 0);
+
+  return (
+    <div className="flex items-center gap-1">
+      {/* Full stars */}
+      {Array.from({ length: fullStars }, (_, i) => (
+        <motion.svg
+          key={`full-${i}`}
+          className="h-4 w-4 text-yellow-400"
+          fill="currentColor"
+          viewBox="0 0 20 20"
+          initial={{ opacity: 0, scale: 0 }}
+          animate={{ opacity: 1, scale: 1 }}
+          transition={{ delay: i * 0.1, type: "spring", stiffness: 200 }}
+        >
+          <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
+        </motion.svg>
+      ))}
+
+      {/* Half star */}
+      {hasHalfStar && (
+        <motion.div
+          className="relative"
+          initial={{ opacity: 0, scale: 0 }}
+          animate={{ opacity: 1, scale: 1 }}
+          transition={{
+            delay: fullStars * 0.1,
+            type: "spring",
+            stiffness: 200,
+          }}
+        >
+          <svg
+            className="h-4 w-4 text-gray-300"
+            fill="currentColor"
+            viewBox="0 0 20 20"
+          >
+            <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
+          </svg>
+          <svg
+            className="absolute top-0 left-0 h-4 w-4 text-yellow-400"
+            fill="currentColor"
+            viewBox="0 0 20 20"
+            style={{ clipPath: "inset(0 50% 0 0)" }}
+          >
+            <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
+          </svg>
+        </motion.div>
+      )}
+
+      {/* Empty stars */}
+      {Array.from({ length: emptyStars }, (_, i) => (
+        <motion.svg
+          key={`empty-${i}`}
+          className="h-4 w-4 text-gray-300"
+          fill="currentColor"
+          viewBox="0 0 20 20"
+          initial={{ opacity: 0, scale: 0 }}
+          animate={{ opacity: 1, scale: 1 }}
+          transition={{
+            delay: (fullStars + (hasHalfStar ? 1 : 0) + i) * 0.1,
+            type: "spring",
+            stiffness: 200,
+          }}
+        >
+          <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
+        </motion.svg>
+      ))}
+
+      <span className="text-muted-foreground ml-1 text-sm">
+        {rating.toFixed(1)}
+      </span>
+    </div>
+  );
+};
