+++ frontend/src/data/constants.ts
@@ -1,0 +1,56 @@
+export const categories = {
+  "Graphics & Design": {
+    subcategories: ["Logo Design", "Illustration", "Web & Mobile Design"],
+    icon: "🎨",
+    color: "from-primary to-primary/70",
+    rating: 4.8,
+    skillsCount: 1247,
+  },
+  "Digital Marketing": {
+    subcategories: [
+      "Social Media Marketing",
+      "Email Marketing",
+      "Video Marketing",
+    ],
+    icon: "📱",
+    color: "from-secondary to-secondary/70",
+    rating: 4.7,
+    skillsCount: 892,
+  },
+  "Writing & Translation": {
+    subcategories: [
+      "Articles & Blog Posts",
+      "Resume Writing",
+      "Technical Writing",
+    ],
+    icon: "✍️",
+    color: "from-accent to-accent/70",
+    rating: 4.6,
+    skillsCount: 634,
+  },
+  "Video & Animation": {
+    subcategories: ["Video Editing", "Short Video Ads", "Intros & Outros"],
+    icon: "🎬",
+    color: "from-destructive to-destructive/70",
+    rating: 4.9,
+    skillsCount: 759,
+  },
+  "Programming & Tech": {
+    subcategories: [
+      "Website Development",
+      "Mobile Development",
+      "Game Development",
+    ],
+    icon: "💻",
+    color: "from-foreground to-foreground/70",
+    rating: 4.8,
+    skillsCount: 1563,
+  },
+  "Data Science & Analytics": {
+    subcategories: ["Data Analysis", "Data Visualization", "Machine Learning"],
+    icon: "📊",
+    color: "from-secondary to-accent",
+    rating: 4.7,
+    skillsCount: 892,
+  },
+};
